Refactor mean reversion strategy and implement endpoints

+++ src/app/schemas/internal/chart_data_dict.py
@@ -3,5 +3,5 @@ from typing import TypedDict
 
 class ChartDataDict(TypedDict):
     date: str
-    equity: int
-    buy_and_hold: int
+    equity: float
+    buy_and_hold: float
